refactor: rename commands and modify structure
