+++ command-snapshot.json
@@ -1,44 +1,44 @@
 [
   {
-    "command": "force:org:shape:create",
+    "command": "org:create:shape",
     "plugin": "@salesforce/plugin-signups",
     "flags": ["api-version", "json", "loglevel", "target-org"],
-    "alias": []
+    "alias": ["force:org:shape:create", "org:shape:create"]
   },
   {
-    "command": "force:org:shape:delete",
-    "plugin": "@salesforce/plugin-signups",
-    "flags": ["api-version", "json", "loglevel", "no-prompt", "target-org"],
-    "alias": []
-  },
-  {
-    "command": "force:org:shape:list",
-    "plugin": "@salesforce/plugin-signups",
-    "flags": ["json", "loglevel", "verbose"],
-    "alias": []
-  },
-  {
-    "command": "force:org:snapshot:create",
+    "command": "org:create:snapshot",
     "plugin": "@salesforce/plugin-signups",
     "flags": ["api-version", "description", "json", "loglevel", "name", "source-org", "target-dev-hub"],
-    "alias": []
+    "alias": ["force:org:snapshot:create", "org:snapshot:create"]
   },
   {
-    "command": "force:org:snapshot:delete",
+    "command": "org:delete:shape",
+    "plugin": "@salesforce/plugin-signups",
+    "flags": ["api-version", "json", "loglevel", "no-prompt", "target-org"],
+    "alias": ["force:org:shape:delete", "org:shape:delete"]
+  },
+  {
+    "command": "org:delete:snapshot",
     "plugin": "@salesforce/plugin-signups",
     "flags": ["api-version", "json", "loglevel", "snapshot", "target-dev-hub"],
-    "alias": []
+    "alias": ["force:org:snapshot:delete", "org:snapshot:delete"]
   },
   {
-    "command": "force:org:snapshot:get",
+    "command": "org:get:snapshot",
     "plugin": "@salesforce/plugin-signups",
     "flags": ["api-version", "json", "loglevel", "snapshot", "target-dev-hub"],
-    "alias": []
+    "alias": ["force:org:snapshot:get", "org:snapshot:get"]
   },
   {
-    "command": "force:org:snapshot:list",
+    "command": "org:list:shape",
+    "plugin": "@salesforce/plugin-signups",
+    "flags": ["json", "loglevel", "verbose"],
+    "alias": ["force:org:shape:list", "org:shape:list"]
+  },
+  {
+    "command": "org:list:snapshot",
     "plugin": "@salesforce/plugin-signups",
     "flags": ["api-version", "json", "loglevel", "target-dev-hub"],
-    "alias": []
+    "alias": ["force:org:snapshot:list", "org:snapshot:list"]
   }
 ]
