+++ src/commands/org/delete/snapshot.ts
@@ -14,7 +14,7 @@ import {
 } from '@salesforce/sf-plugins-core';
 import { Messages } from '@salesforce/core';
 import { SaveResult, SaveError } from 'jsforce';
-import { queryByNameOrId } from '../../../../shared/snapshot';
+import { queryByNameOrId } from '../../../shared/snapshot';
 
 Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('@salesforce/plugin-signups', 'snapshot.delete');
